Observability/tracing-otlp/requirements.txt
openai>=1.60.0
python-dotenv>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
